requirements.txt
flask
numpy
scikit-learn
pillow
